% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addBackquotes}
\alias{addBackquotes}
\title{Add backquotes}
\usage{
addBackquotes(string)
}
\arguments{
\item{string}{character vector}
}
\value{
character vector
}
\description{
(vectorized) Add backquotes when a string has a space in it
}
\examples{
\donttest{
tidyrules:::addBackquotes(c("ab", "a b"))
}

}
\keyword{internal}
