% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rulelist.R
\name{as_rulelist}
\alias{as_rulelist}
\title{as_rulelist generic from \link[=package_tidyrules]{tidyrules} package}
\usage{
as_rulelist(x, ...)
}
\arguments{
\item{x}{object to be coerced to a \link{rulelist}}

\item{...}{for methods to use}
}
\value{
A \link{rulelist}
}
\description{
as_rulelist generic
}
\seealso{
\link{rulelist}, \link{tidy}, \link[=augment.rulelist]{augment}, \link[=predict.rulelist]{predict}

Other Core Rulelist Utility: 
\code{\link{as_rulelist.data.frame}()},
\code{\link{predict.rulelist}()},
\code{\link{print.rulelist}()},
\code{\link{set_keys}()}
}
\concept{Core Rulelist Utility}
