% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruleset.R
\name{as_ruleset}
\alias{as_ruleset}
\title{Get a ruleset from a rulelist}
\usage{
as_ruleset(rulelist)
}
\arguments{
\item{rulelist}{A \link{rulelist}}
}
\value{
A \link{ruleset}
}
\description{
Returns a ruleset object
}
\examples{
model_class_party = partykit::ctree(species ~ .,
                                    data = palmerpenguins::penguins
                                    )
as_ruleset(tidy(model_class_party))

}
\seealso{
\link{rulelist}
}
