% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{augment_class_keys}
\alias{augment_class_keys}
\title{as the name says}
\usage{
augment_class_keys(x, new_data, y_name, weight = 1L, ...)
}
\description{
as the name says
not to be exported
}
\keyword{internal}
