% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rulelist.R
\name{augment_regr_no_keys}
\alias{augment_regr_no_keys}
\title{as the name says}
\usage{
augment_regr_no_keys(x, new_data, y_name, weight, ...)
}
\description{
as the name says
not to be exported
}
\keyword{internal}
