% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rulelist.R
\name{calculate}
\alias{calculate}
\title{\code{calculate} is re-export of \link[generics:calculate]{generics::calculate} from
\link[=package_tidyrules]{tidyrules} package}
\usage{
calculate(x, ...)
}
\arguments{
\item{x}{A \link{rulelist}}

\item{...}{See \link{calculate.rulelist}}
}
\description{
See \link{calculate.rulelist}
}
\seealso{
\link{rulelist}, \link{tidy}, \link[=augment.rulelist]{augment},
\link[=predict.rulelist]{predict}, \link[=calculate.rulelist]{calculate},
\link[=prune.rulelist]{prune}, \link[=reorder.rulelist]{reorder}
}
