% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{package_tidyrules}
\alias{tidyrules}
\alias{tidyrules-package}
\alias{package_tidyrules}
\title{\code{tidyrules}}
\description{
\code{tidyrules} package provides a framework to work with decision
rules. Rules can be extracted from supported models using \link{tidy}, augmented
using validation data by \link[=augment.rulelist]{augment}, manipulated using
standard dataframe operations, (modified) rulelists can be used to
\link[=predict.rulelist]{predict} on unseen (test) data. Utilities include:
Create a rulelist manually (\link[=as_rulelist.data.frame]{as_rulelist}), Export
a rulelist to SQL (\link{to_sql_case}) and so on. The package offers two
classes; \link{rulelist} and \link{ruleset} based on dataframe.
}
\seealso{
\link{rulelist}, \link{tidy}, \link[=augment.rulelist]{augment},
\link[=predict.rulelist]{predict}
}
\author{
\strong{Maintainer}: Srikanth Komala Sheshachala \email{sri.teach@gmail.com}

Authors:
\itemize{
  \item Amith Kumar Ullur Raghavendra \email{amith54@gmail.com}
}

}
