% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rulelist.R
\name{plot.prune_rulelist}
\alias{plot.prune_rulelist}
\title{Plot method for \code{prune_rulelist} class}
\usage{
\method{plot}{prune_rulelist}(x, ...)
}
\arguments{
\item{x}{A 'prune_rulelist' object}

\item{...}{unused}
}
\value{
ggplot2 object (invisibly)
}
\description{
Plot method for \code{prune_rulelist} class
}
