% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{positionSpaceOutsideSinglequotes}
\alias{positionSpaceOutsideSinglequotes}
\title{Position of space outside single quotes}
\usage{
positionSpaceOutsideSinglequotes(string)
}
\arguments{
\item{string}{A chracter vector}
}
\value{
A integer vector of positions
}
\description{
(vectorised) Detect the position of space in a string not within
  a pair of single quotes
}
\examples{
\donttest{
tidyrules:::positionSpaceOutsideSinglequotes(c("hello", "hel' 'o "))
}

}
