% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruleclasses.R
\name{predict.ruleset}
\alias{predict.ruleset}
\title{`predict` method for ruleset class}
\usage{
\method{predict}{ruleset}(object, new_data, raw = FALSE, ...)
}
\arguments{
\item{object}{ruleset object}

\item{new_data}{dataframe to predict}

\item{raw}{(flag, default: FALSE) Whether raw prediction are to be provided}

\item{...}{unused}
}
\value{
A dataframe indicating `rule_nbr` applicable for a `row_nbr` in
  new_data
}
\description{
Returns the `rule_nbr` applicable for a `row_nbr` in new_data
}
\details{
A `row_nbr` is covered more than one `rule_nbr` per 'keys', results in error.

When raw is `FALSE`(default), output is a tidytable/dataframe with three or
more columns: `row_number` (int), columns corresponding to`keys`, `rule_nbr`
(int). If a row number is not covered by any rule, then there is one row with
all other columns other than `row_nbr` has a missing value.

When raw is `TRUE`(default), output is a tidytable/dataframe with three or
more columns: `row_number` (int), columns corresponding to`keys`, `rule_nbr`
(list of integers). If a row number is not covered by any rule, then there is
no row corresponding the `row_nbr`.
}
\examples{
model_rpart = rpart::rpart(species ~ .,
                           data = palmerpenguins::penguins
                           )
tidy_rpart = tidy(model_rpart)
tidy_rpart

output_1 = predict(tidy_rpart, palmerpenguins::penguins)
output_1

output_2 = predict(tidy_rpart, palmerpenguins::penguins, raw = TRUE)
output_2 # `rule_nbr` is a list-column of integer vectors

}
