% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rulelist.R
\name{predict_all_rulelist}
\alias{predict_all_rulelist}
\title{with or without keys}
\usage{
predict_all_rulelist(rulelist, new_data)
}
\arguments{
\item{rulelist}{rulelist}

\item{new_data}{new_data}
}
\value{
dataframe
}
\description{
uses predict_all_nokeys_rulelist
}
\keyword{internal}
