% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruleclasses.R
\name{predict_core}
\alias{predict_core}
\title{Core predict for ruleset/list/tidy set of rules}
\usage{
predict_core(rules_df, new_data)
}
\arguments{
\item{rules_df}{dataframe with at least two columns: `rule_nbr`, `LHS`.
Should have 'keys' columns such that `rule_nbr` along with 'keys' columns
form a unique combo per row}

\item{new_data}{Data to predict on}
}
\value{
dataframe with these columns: `row_nbr` (integer), 'keys' columns,
  `rule_nbr` (list of integers)
}
\description{
Core logic of predict method is written in a generic sense. This
  function will not be exposed at user level.
}
\details{
If a row number is not covered under any rule, then it does not
  appear as a row in the output.
}
\examples{
\dontrun{
library("magrittr")

# ruleset case
rpart::rpart(Species ~ .,data = iris) \%>\%
  tidy() \%>\%
  dplyr::select(rule_nbr, LHS) \%>\%
  predict_core(iris)

# rulelist case
C50::C5.0(species ~.,
          data = palmerpenguins::penguins,
          trials = 5,
          rules = TRUE
          ) \%>\%
  tidy() \%>\%
  dplyr::select(rule_nbr, trial_nbr, LHS) \%>\%
  predict_core(palmerpenguins::penguins)
}
}
