% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rulelist.R
\name{print.prune_rulelist}
\alias{print.prune_rulelist}
\title{Print method for \code{prune_rulelist} class}
\usage{
\method{print}{prune_rulelist}(x, ...)
}
\arguments{
\item{x}{A 'prune_rulelist' object}

\item{...}{unused}
}
\description{
Print method for \code{prune_rulelist} class
}
