% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruleclasses.R
\name{print.rulelist}
\alias{print.rulelist}
\title{Print method for rulelist class}
\usage{
\method{print}{rulelist}(x, ...)
}
\arguments{
\item{x}{A rulelist object}

\item{...}{Passed to `tidytable::print`}
}
\value{
Input (invisibly)
}
\description{
Prints 'keys' and rulelist as a tidytable
}
