% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruleclasses.R
\name{print.ruleset}
\alias{print.ruleset}
\title{Print method for ruleset class}
\usage{
\method{print}{ruleset}(x, ...)
}
\arguments{
\item{x}{A ruleset object}

\item{...}{Passed to `tidytable::print`}
}
\value{
Input (invisibly)
}
\description{
Prints 'keys' and ruleset as a tidytable
}
