% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rulelist.R
\name{prune}
\alias{prune}
\title{\code{prune} is re-export of \link[generics:prune]{generics::prune} from
\link[=package_tidyrules]{tidyrules} package}
\usage{
prune(tree, ...)
}
\arguments{
\item{tree}{A \link{rulelist}}

\item{...}{See \link{prune.rulelist}}
}
\description{
See \link{prune.rulelist}
}
\seealso{
\link{rulelist}, \link{tidy}, \link[=augment.rulelist]{augment},
\link[=predict.rulelist]{predict}, \link[=calculate.rulelist]{calculate},
\link[=prune.rulelist]{prune}, \link[=reorder.rulelist]{reorder}
}
