% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{removeEmptyLines}
\alias{removeEmptyLines}
\title{Remove empty lines}
\usage{
removeEmptyLines(strings)
}
\arguments{
\item{strings}{A character vector}
}
\value{
A character vector
}
\description{
Remove empty strings from a character vector
}
\examples{
\donttest{
tidyrules:::removeEmptyLines(c("abc", "", "d"))
}

}
\keyword{internal}
