% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rulelist.R
\name{reorder}
\alias{reorder}
\title{reorder generic}
\usage{
reorder(x, ...)
}
\arguments{
\item{x}{A \link{rulelist}}

\item{...}{See \link{reorder.rulelist}}
}
\description{
reorder generic for rulelist
}
\seealso{
\link{rulelist}, \link{tidy}, \link[=augment.rulelist]{augment},
\link[=predict.rulelist]{predict}, \link[=calculate.rulelist]{calculate},
\link[=prune.rulelist]{prune}, \link[=reorder.rulelist]{reorder}
}
