% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_translators.R
\name{ruleRToPython}
\alias{ruleRToPython}
\title{Convert a R parsable rule to python parsable rule}
\usage{
ruleRToPython(rule)
}
\arguments{
\item{rule}{(chr vector) R parsable rule(s)}
}
\value{
(chr vector) Python parsable rule(s)
}
\description{
Expected to be passed to `pd.query` method of pandas dataframe
}
