% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_translators.R
\name{ruleRToSQL}
\alias{ruleRToSQL}
\title{Convert a R parsable rule to SQL parsable rule}
\usage{
ruleRToSQL(rule)
}
\arguments{
\item{rule}{(chr vector) R parsable rule(s)}
}
\value{
(chr vector) SQL parsable rule(s) as a 'WHERE' clause
}
\description{
Expected to be passed after SQL 'WHERE' clause
}
