% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{strHead}
\alias{strHead}
\title{Vectorized semantic equivalent of 'head' for a string}
\usage{
strHead(string, n)
}
\arguments{
\item{string}{string}

\item{n}{(integer) Number of characters}
}
\value{
A string
}
\description{
Picks the substring starting from the first character
}
\details{
'n' can be in the interval [-len + 1, len] (both ends inclusive)
}
\examples{
\donttest{
tidyrules:::strHead(c("string", "string2"), 2)
tidyrules:::strHead(c("string", "string2"), -1)
}

}
