% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{strReplaceReduce}
\alias{strReplaceReduce}
\title{Sequential string replace}
\usage{
strReplaceReduce(string, pattern, replacement)
}
\arguments{
\item{string}{string}

\item{pattern}{pattern}

\item{replacement}{replacement}
}
\value{
character vector
}
\description{
Sequential string replace via reduce
}
\examples{
\donttest{
tidyrules:::strReplaceReduce("abcd", c("ab", "dc"), c("cd", "ab"))
}

}
\keyword{internal}
