% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{strSplitSingle}
\alias{strSplitSingle}
\title{String split a string}
\usage{
strSplitSingle(string, pattern)
}
\arguments{
\item{string}{A string}

\item{pattern}{Passed as-is to 'stringr::str_split'}
}
\value{
A character vector
}
\description{
and return a character vector (not a list)
}
\examples{
\donttest{
tidyrules:::strSplitSingle("abc,d", ",")
}

}
