% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{strTail}
\alias{strTail}
\title{Vectorized semantic equivalent of tail for a string}
\usage{
strTail(string, n)
}
\arguments{
\item{string}{string}

\item{n}{(integer) Number of characters}
}
\value{
A string
}
\description{
Picks the substring starting from the first character
}
\details{
'n' can be in the interval [-len + 1, len] (both ends inclusive)
}
\examples{
\donttest{
tidyrules:::strTail(c("string", "string2"), 2)
tidyrules:::strTail(c("string", "string2"), -1)
}

}
