% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c5.R
\name{tidy.C5.0}
\alias{tidy.C5.0}
\title{Obtain rules as rulelist/tiydtable from a C5.0 model}
\usage{
\method{tidy}{C5.0}(x, ...)
}
\arguments{
\item{x}{C5 model fitted with `rules = TRUE`}

\item{...}{Other arguments (See details)}
}
\value{
A rulelist/tidytable where each row corresponds to a rule.
  The columns are: rule_nbr, trial_nbr, LHS, RHS, support, confidence, lift
}
\description{
Each row corresponds to a rule per trial_nbr
}
\details{
Optional named arguments:

\itemize{

\item laplace(flag, default: TRUE) is supported. This computes confidence
with laplace correction as documented under 'Rulesets' here: [C5
doc](https://www.rulequest.com/see5-unix.html).
}
}
\examples{
c5_model = C50::C5.0(Attrition ~., data = modeldata::attrition, rules = TRUE)
summary(c5_model)
tidy(c5_model)
}
