% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/party.R
\name{tidy.constparty}
\alias{tidy.constparty}
\title{Obtain rules as a ruleset/tidytable from a party model}
\usage{
\method{tidy}{constparty}(x, ...)
}
\arguments{
\item{x}{party model}

\item{...}{Other arguments (currently unused)}
}
\value{
A tidytable where each row corresponds to a rule. The columns are:
  rule_nbr, LHS, RHS, support, confidence (for classification only), lift
  (for classification only)
}
\description{
Each row corresponds to a rule. A rule can be copied into
  `dplyr::filter` to filter the observations corresponding to a rule
}
\details{
These party models are supported: regression (y is numeric),
  classification (y is factor)
}
\examples{
model_party_cl = partykit::ctree(species ~ .,data = palmerpenguins::penguins)
model_party_cl
tidy(model_party_cl)

model_party_re = partykit::ctree(bill_length_mm ~ .,
                                 data = palmerpenguins::penguins
                                 )
model_party_re
tidy(model_party_re)
}
