% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubist.R
\name{tidy.cubist}
\alias{tidy.cubist}
\title{Obtain rules as a ruleset/tidytable from a cubist model}
\usage{
\method{tidy}{cubist}(x, ...)
}
\arguments{
\item{x}{Cubist model}

\item{...}{Other arguments (currently unused)}
}
\value{
A ruleset/tidytable where each row corresponds to a rule. The columns
  are: rule_nbr, committee, LHS, RHS, support, mean, min, max, error
}
\description{
Each row corresponds to a rule per committee.
}
\details{
When col_classes argument is missing, an educated guess is made
  about class by parsing the RHS of sub-rule. This might sometimes not lead
  to a parsable rule.
}
\examples{
data("attrition", package = "modeldata")
cols_att = setdiff(colnames(attrition), c("MonthlyIncome", "Attrition"))

cb_att = Cubist::cubist(x = attrition[, cols_att],
                        y = attrition[["MonthlyIncome"]]
                        )
summary(cb_att)
tidy(cb_att)
}
