% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy}
\alias{tidy}
\title{\code{tidy} is re-export of \link[generics:tidy]{generics::tidy} from
\link[=package_tidyrules]{tidyrules} package}
\usage{
tidy(x, ...)
}
\arguments{
\item{x}{A supported model object}

\item{...}{For model specific implementations to use}
}
\description{
\code{tidy} applied on a supported model fit creates a \link{rulelist}.
\strong{See Also} section links to documentation of specific methods.
}
\seealso{
\link{tidy}, \link{tidy.C5.0}, \link{tidy.rpart}, \link{tidy.constparty}, \link{tidy.cubist},
\link{rulelist}, \link[=augment.rulelist]{augment}, \link[=predict.rulelist]{predict}

Other Core Tidy Utility: 
\code{\link{tidy.C5.0}()},
\code{\link{tidy.constparty}()},
\code{\link{tidy.cubist}()},
\code{\link{tidy.rpart}()}
}
\concept{Core Tidy Utility}
