% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpart.R
\name{tidy.rpart}
\alias{tidy.rpart}
\title{Obtain rules as a ruleset/tidytable from a rpart model}
\usage{
\method{tidy}{rpart}(x, ...)
}
\arguments{
\item{x}{rpart model}

\item{...}{Other arguments (currently unused)}
}
\value{
A tidytable where each row corresponds to a rule. The columns are:
  rule_nbr, LHS, RHS, support, confidence (for classification only), lift
  (for classification only)
}
\description{
Each row corresponds to a rule. A rule can be copied into
  `dplyr::filter` to filter the observations corresponding to a rule
}
\details{
NOTE: For rpart rules, one should build the model without
\bold{ordered factor} variable. We recommend you to convert \bold{ordered
factor} to \bold{factor} or \bold{integer} class.
}
\examples{
rpart_class = rpart::rpart(Species ~ .,data = iris)
rpart_class
tidy(rpart_class)

rpart_regr = rpart::rpart(Sepal.Length ~ .,data = iris)
rpart_regr
tidy(rpart_regr)
}
