% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{tidyRules}
\alias{tidyRules}
\title{Obtain rules as a tidy tibble}
\usage{
tidyRules(object, col_classes = NULL, ...)
}
\arguments{
\item{object}{Fitted model object with rules}

\item{col_classes}{Named list or a named character vector of column classes.
Column names of the data used for modeling form the names and the
respective classes for the value. One way of obtaining this is by running
`lapply(data, class)`.}

\item{...}{Other arguments (currently unused)}
}
\value{
A tibble where each row corresponds to a rule
}
\description{
Each row corresponds to a rule. A rule can be copied into
  `dplyr::filter` to filter the observations corresponding to a rule
}
\details{
tidyRule supports these rule based models: C5, Cubist and rpart.
}
\author{
Srikanth KS, \email{sri.teach@gmail.com}
}
