% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpart.R
\name{tidyRules.rpart}
\alias{tidyRules.rpart}
\title{Obtain rules as a tidy tibble from a rpart model}
\usage{
\method{tidyRules}{rpart}(object, ...)
}
\arguments{
\item{object}{Fitted model object with rules}

\item{...}{Other arguments (currently unused)}
}
\value{
A tibble where each row corresponds to a rule. The columns are:
  support, confidence, lift, LHS, RHS
}
\description{
Each row corresponds to a rule. A rule can be copied into
  `dplyr::filter` to filter the observations corresponding to a rule
}
\examples{
data("attrition", package = "rsample")
attrition <- tibble::as_tibble(attrition)
rpart_model <- rpart::rpart(Attrition ~., data = attrition, rules = TRUE)
summary(rpart_model)
tidyRules(rpart_model)
}
\author{
Amith Kumar U R, \email{amith54@gmail.com}
}
