% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rulelist.R
\name{as_rulelist}
\alias{as_rulelist}
\title{as_rulelist generic from \link[=package_tidyrules]{tidyrules} package}
\usage{
as_rulelist(x, ...)
}
\arguments{
\item{x}{object to be coerced to a \link{rulelist}}

\item{...}{for methods to use}
}
\value{
A \link{rulelist}
}
\description{
as_rulelist generic
}
\seealso{
\link{rulelist}, \link{tidy}, \link[=augment.rulelist]{augment},
\link[=predict.rulelist]{predict}, \link[=calculate.rulelist]{calculate},
\link[=prune.rulelist]{prune}, \link[=reorder.rulelist]{reorder}
}
