% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rulelist.R
\name{augment}
\alias{augment}
\title{\code{augment} is re-export of \link[generics:augment]{generics::augment} from
\link[=package_tidyrules]{tidyrules} package}
\usage{
augment(x, ...)
}
\arguments{
\item{x}{A \link{rulelist}}

\item{...}{For methods to use}
}
\description{
See \link{augment.rulelist}
}
\seealso{
\link{rulelist}, \link{tidy}, \link[=augment.rulelist]{augment},
\link[=predict.rulelist]{predict}, \link[=calculate.rulelist]{calculate},
\link[=prune.rulelist]{prune}, \link[=reorder.rulelist]{reorder}
}
