% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_rule_flavor}
\alias{convert_rule_flavor}
\title{Convert a R parsable rule to python/sql parsable rule}
\usage{
convert_rule_flavor(rule, flavor)
}
\arguments{
\item{rule}{(chr vector) R parsable rule(s)}

\item{flavor}{(string) One among: 'python', 'sql'}
}
\value{
(chr vector) of rules
}
\description{
Convert a R parsable rule to python/sql parsable rule
}
\seealso{
\link{rulelist}, \link{tidy}, \link[=augment.rulelist]{augment}, \link[=predict.rulelist]{predict}, \link{to_sql_case}

Other Auxiliary Rulelist Utility: 
\code{\link{to_sql_case}()}
}
\concept{Auxiliary Rulelist Utility}
