% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{package_tidyrules}
\alias{tidyrules}
\alias{tidyrules-package}
\alias{package_tidyrules-package}
\title{About 'tidyrules' package}
\description{
Obtain rules as tidy dataframes
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/talegari/tidyrules}
  \item Report bugs at \url{https://github.com/talegari/tidyrules/issues}
}

}
\author{
\strong{Maintainer}: Srikanth Komala Sheshachala \email{sri.teach@gmail.com}

Authors:
\itemize{
  \item Amith Kumar Ullur Raghavendra \email{amith54@gmail.com}
}

}
