% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rulelist.R
\name{print.rulelist}
\alias{print.rulelist}
\title{Print method for \link{rulelist} class}
\usage{
\method{print}{rulelist}(x, banner = TRUE, ...)
}
\arguments{
\item{x}{A \link{rulelist} object}

\item{banner}{(flag, default: \code{TRUE}) Should the banner be displayed}

\item{...}{Passed to \code{tidytable::print}}
}
\value{
input \link{rulelist} (invisibly)
}
\description{
Prints \link{rulelist} attributes and first few rows.
}
\seealso{
\link{rulelist}, \link{tidy}, \link[=augment.rulelist]{augment},
\link[=predict.rulelist]{predict}, \link[=calculate.rulelist]{calculate},
\link[=prune.rulelist]{prune}, \link[=reorder.rulelist]{reorder}
}
