% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruleset.R
\name{print.ruleset}
\alias{print.ruleset}
\title{Print method for ruleset class}
\usage{
\method{print}{ruleset}(x, banner = TRUE, ...)
}
\arguments{
\item{x}{A \link{rulelist}}

\item{banner}{(flag, default: \code{TRUE}) Should the banner be displayed}

\item{...}{Passed to \code{print.rulelist}}
}
\value{
(invisibly) Returns the ruleset object
}
\description{
Prints the ruleset object
}
\examples{
model_class_party = partykit::ctree(species ~ .,
                                    data = palmerpenguins::penguins
                                    )
as_ruleset(tidy(model_class_party))

}
\seealso{
\link{print.rulelist}
}
